+++ .eslintrc
@@ -51,6 +51,12 @@
         "SwitchCase": 1
       }
     ],
+    "jsx-a11y/label-has-for": [ 2, {
+      "required": {
+          "every": [ "nesting", "id" ]
+      },
+      "allowChildren": true
+    }],
     "react/jsx-boolean-value": "warn",
     "react/jsx-closing-bracket-location": ["error", "line-aligned"],
     "react/jsx-handler-names": [
