+++ client/scripts/Components/__tests__/Checkbox.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import renderer from 'react-test-renderer';
 
-import Checkbox from '../Components/Checkbox';
+import Checkbox from '../Checkbox';
 
 const LABEL = 'LABEL TEXT';
 const mockCallback = jest.fn();
